+++ src/actions/emitters.js
@@ -4,13 +4,15 @@
  * Date: 16.12.2016
  */
 
-export const emitSelected = (e) => ({
+export const emitSelected = (reduxUid, e) => ({
   type: 'EMIT_SELECTED',
+  reduxUid,
   event: e,
 });
 
-export const updateListener = (listener) => ({
+export const updateListener = (reduxUid, listener) => ({
   type: 'UPDATE_LISTENER',
+  reduxUid,
   listener,
 });