+++ src/components/Container.jsx
@@ -8,7 +8,7 @@ import React, { PropTypes } from 'react';
 import classnames from 'classnames';
 import Item from '../containers/Item';
 
-const Container = ({ items, visible }, { classStore }) => (
+const Container = ({ items, visible, reduxStoreName, reduxUid }, { classStore }) => (
   <ul
     className={classnames(
       classStore.classContainer,
@@ -16,7 +16,7 @@ const Container = ({ items, visible }, { classStore }) => (
     )}
   >
     {items.map((item, i) => (
-      <Item key={i} {...item} />
+      <Item key={i} reduxStoreName={reduxStoreName} reduxUid={reduxUid} {...item} />
     ))}
   </ul>
 );
@@ -24,6 +24,8 @@ const Container = ({ items, visible }, { classStore }) => (
 Container.propTypes = {
   items: PropTypes.arrayOf(PropTypes.object).isRequired,
   visible: PropTypes.bool,
+  reduxStoreName: PropTypes.string.isRequired,
+  reduxUid: PropTypes.number.isRequired,
 };
 
 Container.contextTypes = {