+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+
+## [Unreleased]
+### Added
+- `reduxStoreName` prop to customize store field name
+- Multiple `react-metismenu` usage under same Redux `Provider`
+
+## [1.3.0-alpha.2] - 2017-03-01
+### Fixed
+- Some little bugs
+
+## [1.3.0-alpha.1] - 2017-02-03
+### Added
+- `useExternalReduxStore` support for applications with Redux
+
 ## [1.2.2] - 2016-12-20
 ### Fixed
 - `label` and `id` content props now available in custom links like `to`
@@ -168,6 +182,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Development tools
 - Core css file
 
+[Unreleased]: https://github.com/alpertuna/react-metismenu/compare/v1.3.0-alpha.2...master
+[1.3.0-alpha.2]: https://github.com/alpertuna/react-metismenu/compare/v1.3.0-alpha.1...v1.3.0-alpha.2
+[1.3.0-alpha.1]: https://github.com/alpertuna/react-metismenu/compare/v1.2.2...v1.3.0-alpha.1
 [1.2.2]: https://github.com/alpertuna/react-metismenu/compare/v1.2.1...v1.2.2
 [1.2.1]: https://github.com/alpertuna/react-metismenu/compare/v1.1.3...v1.2.1
 [1.1.3]: https://github.com/alpertuna/react-metismenu/compare/v1.1.2...v1.1.3